restructure
- added clean separation for packages
